Update healer.yml
Added more to the notes with information on the character based on stuff from the spread sheet plus a Pros & cons section. I hope the indentation is correct. No clue if or how I add a line between sections of the notes like the character explanation, TLDR and rotation segments.