Update defender.yml
Added Rain. Tried to do his meta data but Rain has some odd talents. His debuff skill is based on a talent, I dunno if you have de-buff classification in the meta data.
He also has a taunt talent which I'm not sure if anyone else has had yet.
I copied Smyu as the template, the HP buff was classed as priority 4 but I thought that was a self buff while the party HP buffs are meant to be party wide buffs applied in combat so I set it as 1 for Rain. But I'm dumb so I'm probably wrong about that.